local Scene

+++ SmartHome/DeviceInfo.h
@@ -69,7 +69,7 @@ enum{
 
 #pragma mark - public
 -(NSData *) author;
--(NSData *)startScenenAtMaster:(int)sceneid;
+-(NSData *)startScenenAtMaster:(long)sceneid;
 
 //TV,DVD,NETV,BGMusic
 -(NSData *) previous:(NSString *)deviceID;
